feat: add videoToAudio and makeVideoList nodes, refactor image nodes
- Add EasyGetAudioFromVideo node to extract audio from video inputs
- Add MakeVideoList node to combine multiple video inputs
- Move MakeRefsCompositeBySam3, SplitImages, MakeImageList to nodes/image.py
- Remove skip_background option from SAM3 composite (breaking change)
- Add ffmpeg_extract_audio utility in utils/video.py
- Reorganize node categories:
  - EasyUse/TimelineEditor: timeline-related nodes
  - EasyUse/MultiTrackEditor: multitrack-related nodes
  - EasyUse/Image: image processing nodes
  - EasyUse/Video: video processing nodes
  - EasyUse/Audio: audio processing nodes
  - EasyUse/Logic: logic nodes (MatchLine)
- Update locales/zh/main.json with new category translations
- Fix test class name EasyVideoToAudio -> EasyGetAudioFromVideo
- Fix logger tag consistency in EasyGetAudioFromVideo
