+++ nodes/__init__.py
@@ -1,4 +1,5 @@
 from .basic import *
+from .image import *
 from .ltxv import *
 from .video import *
 from .wan import *
