+++ tests/test_sam3_refs_composite.py
@@ -1,3 +1,4 @@
+import importlib.util
 import json
 import sys
 import types
@@ -8,8 +9,18 @@ import torch
 from test_multitrack_info_output import _load_basic_module
 
 
+def _load_image_module():
+    _load_basic_module()
+    path = Path(__file__).parents[1] / "nodes" / "image.py"
+    spec = importlib.util.spec_from_file_location("easy_media.nodes.image", path)
+    module = importlib.util.module_from_spec(spec)
+    sys.modules[spec.name] = module
+    spec.loader.exec_module(module)
+    return module
+
+
 def test_make_refs_composite_schema_accepts_lists_and_exposes_expected_widgets():
-    module = _load_basic_module()
+    module = _load_image_module()
 
     schema = module.MakeRefsCompositeBySam3.define_schema()
 
@@ -25,11 +36,9 @@ def test_make_refs_composite_schema_accepts_lists_and_exposes_expected_widgets()
         "detection_threshold",
         "background",
         "composite_mode",
-        "skip_background",
     ]
-    assert schema.inputs[-2].kwargs["options"] == ["original", "sam3_masked"]
-    assert schema.inputs[-2].kwargs["default"] == "original"
-    assert schema.inputs[-1].kwargs["default"] is False
+    assert schema.inputs[-1].kwargs["options"] == ["original", "sam3_masked"]
+    assert schema.inputs[-1].kwargs["default"] == "original"
     assert [output.name for output in schema.outputs] == ["composite", "mask"]
 
 
@@ -49,7 +58,6 @@ def test_make_refs_composite_has_complete_chinese_localization():
         "detection_threshold",
         "background",
         "composite_mode",
-        "skip_background",
     }
     assert translation["inputs"]["background"]["options"] == {
         "white": "白色",
@@ -66,7 +74,7 @@ def test_make_refs_composite_has_complete_chinese_localization():
 
 
 def test_expand_ref_images_splits_a_single_batched_input():
-    module = _load_basic_module()
+    module = _load_image_module()
     batch = torch.stack((torch.zeros(2, 3, 3), torch.ones(2, 3, 3)))
 
     images = module._expand_ref_images([batch])
@@ -78,7 +86,7 @@ def test_expand_ref_images_splits_a_single_batched_input():
 
 
 def test_compose_refs_ignores_empty_masks_and_centers_the_remaining_cutout():
-    module = _load_basic_module()
+    module = _load_image_module()
     ignored = torch.ones(1, 2, 2, 3)
     included = torch.zeros(1, 2, 2, 3)
     included[..., 0] = 1
@@ -102,7 +110,7 @@ def test_compose_refs_ignores_empty_masks_and_centers_the_remaining_cutout():
 
 
 def test_compose_refs_uses_grid_when_full_height_items_do_not_fit():
-    module = _load_basic_module()
+    module = _load_image_module()
     image = torch.ones(1, 2, 4, 3)
     mask = torch.ones(1, 2, 4)
 
@@ -121,9 +129,9 @@ def test_compose_refs_uses_grid_when_full_height_items_do_not_fit():
 
 
 def test_make_refs_composite_runs_sam3_for_each_image_in_a_batch(monkeypatch):
-    module = _load_basic_module()
+    module = _load_image_module()
     calls = []
-    masks = [torch.ones(1, 2, 2)]
+    masks = [torch.ones(1, 2, 2), torch.ones(1, 2, 2)]
 
     class FakeVideoTrack:
         @classmethod
@@ -166,21 +174,20 @@ def test_make_refs_composite_runs_sam3_for_each_image_in_a_batch(monkeypatch):
         detection_threshold=[0.7],
         background=["black"],
         composite_mode=["sam3_masked"],
-        skip_background=[True],
     )
 
     composite, composite_mask = output.values
     assert clip.prompts == ["person"]
-    assert len(calls) == 1
+    assert len(calls) == 2
     assert all(call["detection_threshold"] == 0.7 for call in calls)
     assert all(call["conditioning"] == [({"tokens": "person"}, {})] for call in calls)
     assert composite.shape == (1, 4, 4, 3)
-    assert torch.all(composite == 0)
-    assert torch.all(composite_mask == 1)
+    assert torch.count_nonzero(composite) > 0
+    assert torch.count_nonzero(composite_mask) > 0
 
 
-def test_original_mode_skips_the_last_background_without_running_sam3():
-    module = _load_basic_module()
+def test_original_mode_composites_all_images_without_running_sam3():
+    module = _load_image_module()
 
     class UnusedClip:
         def tokenize(self, prompt):
@@ -202,10 +209,9 @@ def test_original_mode_skips_the_last_background_without_running_sam3():
         detection_threshold=[0.5],
         background=["black"],
         composite_mode=["original"],
-        skip_background=[True],
     )
 
     composite, composite_mask = output.values
-    assert torch.all(composite[..., 0] == 1)
-    assert torch.all(composite[..., 1:] == 0)
-    assert torch.all(composite_mask == 1)
+    assert torch.count_nonzero(composite[..., 0]) > 0
+    assert torch.count_nonzero(composite[..., 1]) > 0
+    assert torch.count_nonzero(composite_mask) > 0
